Add support for incrementally stub programs

+++ stub.bash
@@ -1,5 +1,6 @@
 BATS_MOCK_TMPDIR="${BATS_TMPDIR}"
 BATS_MOCK_BINDIR="${BATS_MOCK_TMPDIR}/bin"
+BATS_MOCK_APPEND=${BATS_MOCK_APPEND-0}
 
 PATH="$BATS_MOCK_BINDIR:$PATH"
 
@@ -15,7 +16,9 @@ stub() {
   mkdir -p "${BATS_MOCK_BINDIR}"
   ln -sf "${BASH_SOURCE[0]%stub.bash}binstub" "${BATS_MOCK_BINDIR}/${program}"
 
-  rm -f "${BATS_MOCK_TMPDIR}/${program}-stub-plan" "${BATS_MOCK_TMPDIR}/${program}-stub-run"
+  if [ $BATS_MOCK_APPEND -ne 1 ]; then {
+    rm -f "${BATS_MOCK_TMPDIR}/${program}-stub-plan" "${BATS_MOCK_TMPDIR}/${program}-stub-run"
+  }; fi
   touch "${BATS_MOCK_TMPDIR}/${program}-stub-plan"
   for arg in "$@"; do printf "%s\n" "$arg" >> "${BATS_MOCK_TMPDIR}/${program}-stub-plan"; done
 }
